Refactor HTTP client classes and rename builder
Replaces AbstractClient with AbstractHttpClient, updating method signatures and timeout handling. Renames ClientBuilder to HttpClientBuilder for clarity and consistency with new class naming conventions.
